requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.3.0